# Change skip* to advance*.