--- a/tests/System.Buffers.ReaderWriter.Tests/Reader_ReadTo.cs
+++ b/tests/System.Buffers.ReaderWriter.Tests/Reader_ReadTo.cs
@@ -3,6 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Buffers.Reader;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 using Xunit;
 
 namespace System.Buffers.Tests
@@ -10,9 +12,11 @@ namespace System.Buffers.Tests
     public class Reader_ReadTo
     {
         [Theory]
-        [InlineData(false)]
-        [InlineData(true)]
-        public void TryReadTo_Span(bool advancePastDelimiter)
+        [InlineData(false, false)]
+        [InlineData(false, true)]
+        [InlineData(true, false)]
+        [InlineData(true, true)]
+        public void TryReadTo_Span(bool advancePastDelimiter, bool useEscapeOverload)
         {
             ReadOnlySequence<byte> bytes = BufferFactory.Create(new byte[][] {
                 new byte[] { 0 },
@@ -22,19 +26,43 @@ namespace System.Buffers.Tests
             });
 
             BufferReader<byte> reader = new BufferReader<byte>(bytes);
+
+            // Read to 0-5
             for (byte i = 0; i < bytes.Length - 1; i++)
             {
                 BufferReader<byte> copy = reader;
-                Assert.True(copy.TryReadTo(out ReadOnlySpan<byte> span, i, 255, advancePastDelimiter));
-                Assert.True(copy.TryReadTo(out span, 6, 255, advancePastDelimiter));
-                Assert.Equal(!advancePastDelimiter, copy.TryReadTo(out span, 6, 255, advancePastDelimiter));
+
+                // Can read to the first integer (0-5)
+                Assert.True(
+                    useEscapeOverload
+                        ? copy.TryReadTo(out ReadOnlySpan<byte> span, i, 255, advancePastDelimiter)
+                        : copy.TryReadTo(out span, i, advancePastDelimiter));
+
+                // Should never have a null Position object
+                Assert.NotNull(copy.Position.GetObject());
+
+                // Should be able to then read to 6
+                Assert.True(
+                    useEscapeOverload
+                        ? copy.TryReadTo(out span, 6, 255, advancePastDelimiter)
+                        : copy.TryReadTo(out span, 6, advancePastDelimiter));
+
+                Assert.NotNull(copy.Position.GetObject());
+
+                // If we didn't advance, we should still be able to read to 6
+                Assert.Equal(!advancePastDelimiter,
+                    useEscapeOverload
+                        ? copy.TryReadTo(out span, 6, 255, advancePastDelimiter)
+                        : copy.TryReadTo(out span, 6, advancePastDelimiter));
             }
         }
 
         [Theory]
-        [InlineData(false)]
-        [InlineData(true)]
-        public void TryReadTo_Sequence(bool advancePastDelimiter)
+        [InlineData(false, false)]
+        [InlineData(false, true)]
+        [InlineData(true, false)]
+        [InlineData(true, true)]
+        public void TryReadTo_Sequence(bool advancePastDelimiter, bool useEscapeOverload)
         {
             ReadOnlySequence<byte> bytes = BufferFactory.Create(new byte[][] {
                 new byte[] { 0 },
@@ -44,12 +72,38 @@ namespace System.Buffers.Tests
             });
 
             BufferReader<byte> reader = new BufferReader<byte>(bytes);
+
+            // Read to 0-5
             for (byte i = 0; i < bytes.Length - 1; i++)
             {
                 BufferReader<byte> copy = reader;
-                Assert.True(copy.TryReadTo(out ReadOnlySequence<byte> span, i, 255, advancePastDelimiter));
-                Assert.True(copy.TryReadTo(out span, 6, 255, advancePastDelimiter));
-                Assert.Equal(!advancePastDelimiter, copy.TryReadTo(out span, 6, 255, advancePastDelimiter));
+
+                // Can read to the first integer (0-5)
+                Assert.True(
+                    useEscapeOverload
+                        ? copy.TryReadTo(out ReadOnlySequence<byte> sequence, i, 255, advancePastDelimiter)
+                        : copy.TryReadTo(out sequence, i, advancePastDelimiter));
+
+                // Should never have a null Position object
+                Assert.NotNull(copy.Position.GetObject());
+                var enumerator = sequence.GetEnumerator();
+                while (enumerator.MoveNext()) ;
+
+                // Should be able to read to final 6
+                Assert.True(
+                    useEscapeOverload
+                        ? copy.TryReadTo(out sequence, 6, 255, advancePastDelimiter)
+                        : copy.TryReadTo(out sequence, 6, advancePastDelimiter));
+
+                Assert.NotNull(copy.Position.GetObject());
+                enumerator = sequence.GetEnumerator();
+                while (enumerator.MoveNext()) ;
+
+                // If we didn't advance, we should still be able to read to 6
+                Assert.Equal(!advancePastDelimiter,
+                    useEscapeOverload
+                        ? copy.TryReadTo(out sequence, 6, 255, advancePastDelimiter)
+                        : copy.TryReadTo(out sequence, 6, advancePastDelimiter));
             }
         }