Fix moving past empty sequences.
Fix span advance to advancePastDelimiter handling.
Add IsLastSegment.
Add 4x faster Position property.
Add a variety of tests.