@JeremyKuhne 2018-10-12
A faster Position showed 5-7% benefits in other benchmarks (such as TryReadUntil). On it's own it is much faster:

**Before:**
```
          Method |     Mean |    Error |   StdDev |
---------------- |---------:|---------:|---------:|
 Reader_Position | 106.9 us | 1.259 us | 1.051 us |
```
**After:**
```
          Method |     Mean |     Error |    StdDev |
---------------- |---------:|----------:|----------:|
 Reader_Position | 24.84 us | 0.4780 us | 0.5870 us |
```